requirements.txt
aiofiles==22.1.0
alembic==1.9.4
greenlet==2.0.1
httptools==0.5.0
Mako==1.2.4
MarkupSafe==2.1.2
multidict==6.0.4
packaging==23.0
psycopg2-binary==2.9.5
pydantic==1.10.4
PyJWT==2.6.0
python-dotenv==0.21.0
PyYAML==6.0
sanic==22.12.0
Sanic-Cors==2.2.0
sanic-openapi==21.12.0
sanic-routing==22.8.0
SQLAlchemy==1.4.46
typing_extensions==4.4.0
websockets==10.4
